Create README.md
Implementation of an API Mashup involving the following two API's
-RottenTomatoes Box Office Movies API(http://developer.rottentomatoes.com/docs/read/json/v10/Box_Office_Movies)
-GraceNote OnConnect Movie Showtimes API(http://developer.tmsapi.com/docs/data_v1/movies/Movie_showtimes)

The function of this application is to list the top 10 movies int the Box Office according the RottenTomatoes. On clicking a movie shows the showtimes(within US and Canada) for that movie within 5 miles of your location for that day.